@@ -6,8 +6,7 @@ namespace Laminas\ApiTools\Doctrine\Admin\Controller;
 
 use Interop\Container\ContainerInterface;
 use Laminas\ApiTools\Doctrine\Admin\Model\DoctrineAutodiscoveryModel;
-use Laminas\ServiceManager\AbstractPluginManager;
-use Laminas\ServiceManager\FactoryInterface;
+use Laminas\ServiceManager\Factory\FactoryInterface;
 use Laminas\ServiceManager\ServiceLocatorInterface;
 
 class DoctrineAutodiscoveryControllerFactory implements FactoryInterface
@@ -36,10 +35,6 @@ class DoctrineAutodiscoveryControllerFactory implements FactoryInterface
      */
     public function createService(ServiceLocatorInterface $container)
     {
-        if ($container instanceof AbstractPluginManager) {
-            $container = $container->getServiceLocator() ?: $container;
-        }
-
         return $this($container, DoctrineAutodiscoveryController::class);
     }
 }
