@@ -4,12 +4,13 @@ declare(strict_types=1);
 
 namespace Laminas\ApiTools\Doctrine\Admin\Model;
 
-use Doctrine\Common\Persistence\Mapping\AbstractClassMetadataFactory;
+use Doctrine\Persistence\Mapping\AbstractClassMetadataFactory;
 use Exception;
 use Laminas\ApiTools\Admin\Model\RestServiceEntity;
 use Laminas\ApiTools\ApiProblem\ApiProblem;
 use Laminas\ApiTools\Rest\AbstractResourceListener;
 use Laminas\ServiceManager\ServiceManager;
+use Laminas\Stdlib\Parameters;
 
 class DoctrineMetadataServiceResource extends AbstractResourceListener
 {
@@ -74,7 +75,7 @@ class DoctrineMetadataServiceResource extends AbstractResourceListener
     /**
      * Fetch metadata for all REST services
      *
-     * @param array $params
+     * @param Parameters|array $params
      * @return RestServiceEntity[]|ApiProblem
      */
     public function fetchAll($params = [])
