@@ -4,7 +4,7 @@ declare(strict_types=1);
 
 namespace Laminas\ApiTools\Doctrine\Admin\Model;
 
-use Doctrine\Common\Persistence\ObjectManager;
+use Doctrine\Persistence\ObjectManager;
 use Laminas\ApiTools\Admin\Model\NewRestServiceEntity as LaminasNewRestServiceEntity;
 use Laminas\Stdlib\ArraySerializableInterface;
 
@@ -32,6 +32,10 @@ class NewDoctrineServiceEntity extends LaminasNewRestServiceEntity implements Ar
     {
         parent::exchangeArray($data);
 
+        /**
+         * @var string $key
+         * @var mixed $value
+         */
         foreach ($data as $key => $value) {
             $key = strtolower($key);
             $key = str_replace('_', '', $key);
