@@ -52,6 +52,6 @@ class Module
      */
     public function getModuleDependencies()
     {
-        return ['Phpro\DoctrineHydrationModule'];
+        return ['ApiSkeletons\DoctrineORMHydrationModule'];
     }
 }
