@@ -4,6 +4,7 @@ declare(strict_types=1);
 
 namespace Laminas\ApiTools\Doctrine\Server\Query\Provider;
 
+use Doctrine\ORM\QueryBuilder;
 use Laminas\ApiTools\Rest\ResourceEvent;
 
 class DefaultOrm extends AbstractQueryProvider
@@ -11,10 +12,10 @@ class DefaultOrm extends AbstractQueryProvider
     /**
      * @param string $entityClass
      * @param array $parameters
-     * @return mixed This will return an ORM or ODM Query\Builder
      */
-    public function createQuery(ResourceEvent $event, $entityClass, $parameters)
+    public function createQuery(ResourceEvent $event, $entityClass, $parameters): QueryBuilder
     {
+        /** @var QueryBuilder $queryBuilder */
         $queryBuilder = $this->getObjectManager()->createQueryBuilder();
         $queryBuilder
             ->select('row')
