@@ -6,7 +6,7 @@ return [
     'modules'                 => [
         'DoctrineModule',
         'DoctrineMongoODMModule',
-        'Phpro\DoctrineHydrationModule',
+        'ApiSkeletons\DoctrineORMHydrationModule',
         'Laminas\ApiTools',
         'Laminas\ApiTools\Admin',
         'Laminas\ApiTools\Hal',
