@@ -8,7 +8,7 @@ use Interop\Container\ContainerInterface;
 use Laminas\ApiTools\Doctrine\Admin\Controller\DoctrineAutodiscoveryController;
 use Laminas\ApiTools\Doctrine\Admin\Controller\DoctrineAutodiscoveryControllerFactory;
 use Laminas\ApiTools\Doctrine\Admin\Model\DoctrineAutodiscoveryModel;
-use Laminas\ServiceManager\AbstractPluginManager;
+use Laminas\ServiceManager\ServiceLocatorInterface;
 use LaminasTest\ApiTools\Doctrine\DeprecatedAssertionsTrait;
 use PHPUnit\Framework\TestCase;
 use Prophecy\PhpUnit\ProphecyTrait;
@@ -45,11 +45,11 @@ class DoctrineAutodiscoveryControllerFactoryTest extends TestCase
 
     public function testLegacyFactoryReturnsDoctrineAutodiscoveryController(): void
     {
-        $controllers = $this->prophesize(AbstractPluginManager::class);
-        $controllers->getServiceLocator()->willReturn($this->container->reveal());
+        $this->container = $this->prophesize(ServiceLocatorInterface::class);
+        $this->container->get(DoctrineAutodiscoveryModel::class)->willReturn($this->model);
 
         $factory    = new DoctrineAutodiscoveryControllerFactory();
-        $controller = $factory->createService($controllers->reveal());
+        $controller = $factory->createService($this->container->reveal());
 
         $this->assertInstanceOf(DoctrineAutodiscoveryController::class, $controller);
         $this->assertAttributeSame($this->model, 'model', $controller);
