@@ -4,14 +4,14 @@ declare(strict_types=1);
 
 namespace LaminasTest\ApiTools\Doctrine\Server\Event\Listener;
 
-use Doctrine\Common\Persistence\ObjectManager;
 use Doctrine\Instantiator\InstantiatorInterface;
+use Doctrine\ORM\EntityManager;
 use Doctrine\ORM\Mapping\ClassMetadata;
 use Laminas\ApiTools\Doctrine\Server\Event\Listener\CollectionListener;
 use Laminas\Hydrator\HydratorInterface;
 use LaminasTestApiToolsDb\Entity\Artist;
+use PHPUnit\Framework\MockObject\MockObject;
 use PHPUnit\Framework\TestCase;
-use PHPUnit_Framework_MockObject_MockObject;
 use ReflectionMethod;
 use ReflectionProperty;
 
@@ -27,11 +27,11 @@ class CollectionListenerTest extends TestCase
         $artist = $this->getMockBuilder(Artist::class)->getMock();
         $data   = [];
 
-        /** @var ObjectManager|PHPUnit_Framework_MockObject_MockObject $om */
-        $om            = $this->getMockBuilder(ObjectManager::class)->getMock();
+        /** @var EntityManager&MockObject $om */
+        $om            = $this->createMock(EntityManager::class);
         $classMetadata = $this->getMockBuilder(ClassMetadata::class)
-                ->disableOriginalConstructor()
-                ->getMock();
+            ->disableOriginalConstructor()
+            ->getMock();
         $classMetadata->expects(self::once())
             ->method('getIdentifierFieldNames')
             ->with(Artist::class)
@@ -52,8 +52,8 @@ class CollectionListenerTest extends TestCase
             ->with($data, self::isInstanceOf(Artist::class));
 
         if ($withEntityFactory) {
-            /** @var InstantiatorInterface|PHPUnit_Framework_MockObject_MockObject $entityFactory */
-            $entityFactory = $this->getMockBuilder(InstantiatorInterface::class)->getMock();
+            /** @var InstantiatorInterface&MockObject $entityFactory */
+            $entityFactory = $this->createMock(InstantiatorInterface::class);
 
             $entityFactory->expects(self::once())
                 ->method('instantiate')
@@ -66,7 +66,10 @@ class CollectionListenerTest extends TestCase
         $listener = new CollectionListener($entityFactory);
         $listener->setObjectManager($om);
 
-        $hydratorMapProperty = new ReflectionProperty($listener, 'entityHydratorMap');
+        $hydratorMapProperty = new ReflectionProperty(
+            $listener,
+            'entityHydratorMap'
+        );
         $hydratorMapProperty->setAccessible(true);
         $hydratorMapProperty->setValue($listener, [Artist::class => $hydrator]);
 
